requirements.txt
# Core dependencies
streamlit>=1.31.0
python-dotenv>=1.0.0
requests>=2.31.0
# pydantic>=2.6.0
pydantic>=2.0.0,<3.0.0


openai>=1.12.0
# AI/ML dependencies
# crewai>=0.11.0
crewai==0.108.0 
# crewai-tools==0.108.0
crewai-tools==0.38.1

langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.13



# Document processing
groundx>=0.1.0

# Optional: For PDF processing (if needed)
PyPDF2>=3.0.0

sentence-transformers>=2.2.2

# Optional: For PDF processing (if needed)
# Core dependencies
# ChromaDB and its dependencies with specific versions
# chromadb==0.4.18  # Try an earlier stable version
chromadb>=0.5.23
onnxruntime==1.16.3

# Additional dependencies that might be needed
numpy==1.24.3
# pypika<0.48.9
PyPDF2==3.0.1
# langchain==0.1.0
# langchain-community==0.0.13
tiktoken==0.7.0

# System compatibility
pytesseract==0.3.10
Pillow==10.1.0


pysqlite3-binary>=0.5.0